split message api endpoints
