+++ backend/oasst_backend/api/v1/api.py
@@ -1,8 +1,11 @@
 # -*- coding: utf-8 -*-
 from fastapi import APIRouter
-from oasst_backend.api.v1 import management, tasks, text_labels
+from oasst_backend.api.v1 import frontend_messages, frontend_users, messages, tasks, text_labels, users
 
 api_router = APIRouter()
 api_router.include_router(tasks.router, prefix="/tasks", tags=["tasks"])
 api_router.include_router(text_labels.router, prefix="/text_labels", tags=["text_labels"])
-api_router.include_router(management.router, prefix="/management", tags=["management"])
+api_router.include_router(messages.router, prefix="/messages", tags=["messages"])
+api_router.include_router(frontend_messages.router, prefix="/frontend_messages", tags=["frontend_messages"])
+api_router.include_router(users.router, prefix="/users", tags=["users"])
+api_router.include_router(frontend_users.router, prefix="/frontend_users", tags=["frontend_users"])
